various refactoring
